gh-353  regrouped the views
