+++ Multisig/UI/Transactions/AdvancedTransactionDetailsView.swift
