+++ Multisig/UI/Transactions/TransactionDetails/Header Views/ChangeMastercopyTransactionDetailsHeaderView.swift
