+++ Multisig/UI/Transactions/TransactionDetails/TransactionDetailsView+ConfirmationsView/TransactionConfirmationViewModel.swift
